+++ app/src/main/kotlin/com/simplemobiletools/calendar/models/CalDAVCalendar.kt
@@ -1,5 +1,5 @@
 package com.simplemobiletools.calendar.models
 
-data class CalDAVCalendar(val id: Long, val displayName: String, val accountName: String, val ownerName: String, val color: Int, val accessLevel: Int) {
+data class CalDAVCalendar(val id: Int, val displayName: String, val accountName: String, val ownerName: String, val color: Int, val accessLevel: Int) {
     fun canWrite() = accessLevel >= 500
 }
